Frame colouring is based on angle
Colour on scale from 0 to 1.
Molecules that are antiparallel will be the same colour

Also increasing default key frame rate

+++ lib/mol_functions.h
@@ -21,7 +21,7 @@ static double STRUCT_DIST = 0.3;
 vect orientation(molecule *, Frame *);
 double angle(molecule *, Frame *);
 vect wrap_x(vect v, double a);
-int mol_colour(molecule * m, Frame * frame);
+double mol_colour(molecule * m, Frame * frame);
 double struct_relax(molecule * m, Frame * frame);
 double hexatic(int n, molecule* m1, Frame *frame);
 double circle_ordering(molecule *m);
@@ -29,4 +29,3 @@ int circle_colour(molecule * m);
 int neighbour_colour(molecule * m, Frame *frame);
 
 #endif /* defined(MY_MOL_FUNCTIONS) */
-
